eliminate gbfs code
Summary: This PR eliminates the gbfs code because it turns out the real time feed does not connect with the trip data at all. However, we have paths forward to demo the bay-bikes data with daily, monthly, and scheduled retraining pipelines so we will be able to effectively test out the partitions and backfill systems.

Test Plan: unit

Reviewers:

Subscribers: